Move server test for context images in engine/tests
